% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{add_threads}
\alias{add_threads}
\title{Add thread information to a data frame}
\usage{
add_threads(df)
}
\arguments{
\item{df}{a tibble created by download_sub_urls() or download_keyword_urls()}
}
\value{
a modified tibble, with two extra list columns: nodes and edges
}
\description{
Add thread information to a data frame
}
