% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{auc.rawrrChromatogram}
\alias{auc.rawrrChromatogram}
\title{deriving area under the curve (AUC)}
\usage{
auc.rawrrChromatogram(x)
}
\arguments{
\item{x}{an rawrrChromatogram object contains \code{x$times} and
\code{x$intensities}. \code{x$times} is assumed to be in minutes.}
}
\value{
A numeric value.
}
\description{
deriving area under the curve (AUC)
}
