% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFunctions.R
\name{basePeak}
\alias{basePeak}
\title{Base peak of a spectrum}
\usage{
basePeak(x)
}
\arguments{
\item{x}{A rawrrSpectrum object}
}
\value{
A double vector of length two. The first component is the base peak
position (m/z). The second component is the base peak intensity.
}
\description{
Base peak of a spectrum
}
\examples{
S <- readSpectrum(rawfile = sampleFilePath(), 1)
basePeak(S[[1]])
}
