% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff_net.R
\name{create_aff_net}
\alias{create_aff_net}
\title{Affiliation network}
\usage{
create_aff_net(user_list)
}
\arguments{
\item{user_list}{a list of unique users}
}
\value{
A named list. Names are users and values are subreddits.
}
\description{
Affiliation network
}
