% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{dependentScan}
\alias{dependentScan}
\title{Retrieve dependent scan(s) of a scan listed in scan index}
\usage{
dependentScan(x, scanNumber)
}
\arguments{
\item{x}{A scan index returned by \code{readIndex}.}

\item{scanNumber}{The scan number that should be inspected for dependent scans.}
}
\value{
The scan number of the dependent scan(s).
}
\description{
Retrieve dependent scan(s) of a scan listed in scan index
}
\examples{
Idx <- readIndex(rawfile = sampleFilePath())
dependentScan(Idx, scanNumber = 1)
}
