% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{.addNupkgSource}
\alias{.addNupkgSource}
\title{dotnet nuget add source /Users/cp/Library/Caches/org.R-project.R/R/rawrr/rawrrassembly/
dotnet nuget remove source "Package source 1"
dotnet nuget list source}
\usage{
.addNupkgSource()
}
\description{
dotnet nuget add source /Users/cp/Library/Caches/org.R-project.R/R/rawrr/rawrrassembly/
dotnet nuget remove source "Package source 1"
dotnet nuget list source
}
