% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{.addPackages}
\alias{.addPackages}
\title{dotnet add package ThermoFisher.CommonCore.MassPrecisionEstimator}
\usage{
.addPackages(dir, version = "8.0.6")
}
\description{
dotnet add package ThermoFisher.CommonCore.MassPrecisionEstimator
}
