% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{.checkDllInMonoPath}
\alias{.checkDllInMonoPath}
\title{Check if a file is contained in the environment variable \code{MONO_PATH}.}
\usage{
.checkDllInMonoPath(dll = "ThermoFisher.CommonCore.Data.dll")
}
\arguments{
\item{dll}{a file name.}
}
\value{
a boolean
}
\description{
Check if a file is contained in the environment variable \code{MONO_PATH}.
}
