% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{.downloadNupkgs}
\alias{.downloadNupkgs}
\alias{Thermo}
\alias{ThermoFisher}
\alias{ThermoFisherScientific}
\title{Download and install the Thermo Fisher Scientific .NET 8.0 nupkgs}
\usage{
.downloadNupkgs(sourceUrl = .thermofisherlsmsUrl(), force = TRUE)
}
\arguments{
\item{sourceUrl}{url of nupkgs.}

\item{force}{if \code{TRUE} it will overwrite the pkgs.}
}
\value{
An (invisible) vector of integer code, 0 for success and non-zero for
failure. For the "wget" and "curl" methods this is the status code returned
by the external program.
}
\description{
Download and install the Thermo Fisher Scientific .NET 8.0 nupkgs
}
\details{
The console application assembly \code{rawrr.exe} requires:
\itemize{
\item {\code{ThermoFisher.CommonCore.Data.dll}, }
\item{\code{ThermoFisher.CommonCore.MassPrecisionEstimator.dll}, and}
\item{ThermoFisher.CommonCore.RawFileReader.dll}
}.
}
\references{
\itemize{
  \item{\url{https://www.mono-project.com/docs/advanced/assemblies-and-the-gac/}}
  \item{\url{https://planetorbitrap.com/rawfilereader}}
  \item{\doi{10.1021/acs.jproteome.0c00866}}
}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>, 2021, 2024
}
