% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.plotLockMassCorrection}
\alias{.plotLockMassCorrection}
\title{lock mass correction plot}
\usage{
.plotLockMassCorrection(rawfile)
}
\arguments{
\item{rawfile}{the name of the raw file containing the mass spectrometry data from the Thermo Fisher Scientific instrument.}
}
\value{
a ggplot object
}
\description{
lock mass correction plot
}
\examples{
rawrr::sampleFilePath() |> rawrr:::.plotLockMassCorrection()
}
\references{
rawDiag \doi{10.1021/acs.jproteome.8b00173}
}
\author{
Christian Trachsel (2017), Christian Panse (2023)
}
