% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{.thermofisherlsmsUrl}
\alias{.thermofisherlsmsUrl}
\title{URL for Thermo Fisher .NET assemblies}
\usage{
.thermofisherlsmsUrl()
}
\value{
an URL
}
\description{
URL for Thermo Fisher .NET assemblies
}
