% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{filter}
\alias{filter}
\title{derive scans number which match a specified filter}
\usage{
filter(rawfile, filter = "ms", precision = 10, tmpdir = tempdir())
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scietific raw file}

\item{filter}{scan filter string, e.g., \code{ms} or \code{ms2}}

\item{precision}{mass precision, default is 10.}

\item{tmpdir}{defines the directory used to store temporary data generated by
the .NET assembly \code{rawrr.exe}. The default uses \code{tempdir()}.}
}
\value{
a vecntor of integer values.
}
\description{
derive scans number which match a specified filter
}
