% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praw.R
\name{init_reddit}
\alias{init_reddit}
\title{Initialize Reddit instance}
\usage{
init_reddit(username, password)
}
\arguments{
\item{username}{username}

\item{password}{password}
}
\description{
Initialize Reddit instance
}
