% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praw.R
\name{install_praw}
\alias{install_praw}
\title{Import PRAW}
\usage{
install_praw(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a method that will work in the local environment. Change the default to force a specific installation method. Note that the "virtualenv" method is not available on Windows.}

\item{conda}{The path to a conda executable. Use "auto" to allow reticulate to automatically find an appropriate conda binary.}
}
\description{
This a simple wrapper for the reticulate::py_install() function.
}
