% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{installRawFileReaderDLLs}
\alias{installRawFileReaderDLLs}
\alias{Thermo}
\alias{ThermoFisher}
\alias{ThermoFisherScientific}
\title{Download and install the New RawFileReader from Thermo Fisher Scientific .Net
assemblies}
\usage{
installRawFileReaderDLLs(sourceUrl = "https://tinyurl.com/4d3awfyv/", ...)
}
\arguments{
\item{sourceUrl}{url of New RawFileReader from Thermo Fisher Scientific
assemblies.}

\item{...}{other parameter for \code{download.file}}
}
\value{
An (invisible) vector of integer code, 0 for success and non-zero for
failure. For the "wget" and "curl" methods this is the status code returned
by the external program.
}
\description{
Download and install the New RawFileReader from Thermo Fisher Scientific .Net
assemblies
}
\details{
The console application assembly \code{rawrr.exe} requires three
assemplies:
\itemize{
\item {\code{ThermoFisher.CommonCore.Data.dll}, }
\item{\code{ThermoFisher.CommonCore.MassPrecisionEstimator.dll}, and}
\item{ThermoFisher.CommonCore.RawFileReader.dll}
}.

The \code{rawrr.exe} assembly can be built from C# source code by using the
\code{msbuild} tool shipped by the \url{https://www.mono-project.com} or by
Microsoft's .NET SDK \url{https://dotnet.microsoft.com} on Linux, Microsoft,
and macOS.

If no build tool and C# compiler (\code{csc} or \code{msc}) are available or
the build process fails, you can download \code{rawrr.exe} assembly from the
authors' site.
}
\examples{
# to install all assemblies
\donttest{
rawrr::installRawFileReaderDLLs() 
rawrr:::buildRawrrExe() || rawrr::installRawrrExe()
}
}
\references{
\itemize{
  \item{\url{https://www.mono-project.com/docs/advanced/assemblies-and-the-gac/}}
  \item{\url{https://planetorbitrap.com/rawfilereader}}
  \item{\url{https://doi.org/10.1021/acs.jproteome.0c00866}}
}
}
\seealso{
\link{buildRawrrExe} and \link{installRawrrExe}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>, 2021
}
