% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{installRawrrExe}
\alias{installRawrrExe}
\alias{rawrr.exe}
\title{Download \code{rawrr} assembly}
\usage{
installRawrrExe(
  sourceUrl = "https://fgcz-ms.uzh.ch/~cpanse/rawrr/dotnet/1.17.1/",
  force = FALSE,
  ...
)
}
\arguments{
\item{sourceUrl}{url of \code{rawrr.exe} assembly.}

\item{force}{if \code{TRUE} it will overwrite the assembly}

\item{...}{other parameter for \code{download.file}.}
}
\value{
An integer code, 0 for success and non-zero for
failure. For the "wget" and "curl" methods this is the status code returned
by the external program.
}
\description{
downloads and installs the \code{rawrr.exe} .NET assembly in
the directory provided by \code{rawrrAssemblyPath()}.
}
\details{
The console application \code{rawrr} is used by the package's
reader functions through a \link{system2} call.
}
\seealso{
\link{buildRawrrExe}
}
