% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{is.rawrrChromatogram}
\alias{is.rawrrChromatogram}
\title{Check if object is instance of class \code{rawrrChromatogram}}
\usage{
is.rawrrChromatogram(x)
}
\arguments{
\item{x}{The object to be tested.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if object is instance of class \code{rawrrChromatogram}
}
\examples{
rawfile <- sampleFilePath()
C <- readChromatogram(rawfile, mass = 445.1181, tol = 10)
is.rawrrChromatogram(C[[1]])
}
\author{
Tobias Kockmann, 2020.
}
