% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{is.rawrrSpectrum}
\alias{is.rawrrSpectrum}
\title{Check if object is instance of class \code{rawrrSpectrum}}
\usage{
is.rawrrSpectrum(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if object is instance of class \code{rawrrSpectrum}
}
\examples{
rawfile <- sampleFilePath()
S <- readSpectrum(rawfile, scan = 1:10)
is.rawrrSpectrum(S[[1]])
}
