% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{is.rawrrSpectrumSet}
\alias{is.rawrrSpectrumSet}
\title{Function to check if an object is an instance of class \code{rawrrSpectrumSet}}
\usage{
is.rawrrSpectrumSet(x)
}
\arguments{
\item{x}{any R object to be tested.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Function to check if an object is an instance of class \code{rawrrSpectrumSet}
}
\examples{
rawfile <- sampleFilePath()
S <- readSpectrum(rawfile, scan = 1:10)
is.rawrrSpectrum(S[[1]])
}
