% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{is.rawRspectrum}
\alias{is.rawRspectrum}
\title{Test if object is instance of class \code{rawRspectrum}}
\usage{
is.rawRspectrum(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
TRUE or FALSE
}
\description{
Test if object is instance of class \code{rawRspectrum}
}
\examples{

pathToRawFile <- file.path(path.package(package = 'rawR'), 'extdata', 'sample.raw')
S <- readSpectrum(pathToRawFile, scans = 1:10)
is.rawRspectrum(S[[1]])
}
