% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFunctions.R
\name{massRange}
\alias{massRange}
\title{Acquisition/scan range of spectrum}
\usage{
massRange(x)
}
\arguments{
\item{x}{A rawrrSpectrum object}
}
\value{
A double vector of length two. The first component is the start m/z, the
second is the stop m/z value used by the detector during data acquisition.
Also referred to as scan range.
}
\description{
Acquisition/scan range of spectrum
}
\examples{
S <- readSpectrum(rawfile = sampleFilePath(), 1)
massRange(S[[1]])
}
