% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{masterScan}
\alias{masterScan}
\title{Retrieve master scan of scan listed in scan index}
\usage{
masterScan(x, scanNumber)
}
\arguments{
\item{x}{A scan index returned by \code{readIndex}.}

\item{scanNumber}{The scan number that should be inspected for the presence
of a master scan.}
}
\value{
Returns the scan number of the master scan or NA if no master scan
exists.
}
\description{
Retrieve master scan of scan listed in scan index
}
\examples{
Idx <- readIndex(rawfile = sampleFilePath())
masterScan(Idx, scanNumber = 1)
}
