% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{new_rawrrSpectrum}
\alias{new_rawrrSpectrum}
\title{Create instances of class \code{rawrrSpectrum}}
\usage{
new_rawrrSpectrum(
  scan = numeric(),
  massRange = numeric(),
  scanType = character(),
  rtinseconds = numeric(),
  centroidStream = logical(),
  mZ = numeric(),
  intensity = numeric()
)
}
\arguments{
\item{scan}{scan number}

\item{massRange}{Mass range covered by spectrum}

\item{scanType}{Character string describing the scan type.}

\item{rtinseconds}{Retention time in seconds}

\item{centroidStream}{Logical indicating if centroided data is available}

\item{mZ}{m/z values}

\item{intensity}{Intensity values}
}
\value{
Object of class \code{rawrrSpectrum}
}
\description{
Developer function.
}
\author{
Tobias Kockmann, 2020.
}
