% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{plot.rawRchromatogram}
\alias{plot.rawRchromatogram}
\title{Plot \code{rawRchromatogram} objects}
\usage{
\method{plot}{rawRchromatogram}(x, legend = TRUE, ...)
}
\arguments{
\item{x}{A \code{rawRchromatogram} object to be plotted.}

\item{legend}{Should legend be printed?}

\item{...}{Passes additional arguments.}
}
\description{
Plot \code{rawRchromatogram} objects
}
\examples{
rawfile <- sample()
C <- readChromatogram(rawfile, mass = 445.1181, tol = 10)
plot(C[[1]])
}
