% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{plot.rawRchromatogramSet}
\alias{plot.rawRchromatogramSet}
\title{Plot \code{rawRchromatogramSet} objects}
\usage{
\method{plot}{rawRchromatogramSet}(x, diagnostic = FALSE, ...)
}
\arguments{
\item{x}{A \code{rawRchromatogramSet} object to be plotted.}

\item{diagnostic}{Show diagnostic legend?}

\item{...}{Passes additional arguments.}
}
\description{
Plot \code{rawRchromatogramSet} objects
}
\author{
Tobias Kockmann, 2020.
}
