% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{plot.rawrrChromatogram}
\alias{plot.rawrrChromatogram}
\title{Plot \code{rawrrChromatogram} objects}
\usage{
\method{plot}{rawrrChromatogram}(x, legend = TRUE, ...)
}
\arguments{
\item{x}{A \code{rawrrChromatogram} object to be plotted.}

\item{legend}{Should legend be printed?}

\item{...}{Passes additional arguments.}
}
\description{
Plot \code{rawrrChromatogram} objects
}
\examples{
rawfile <- sampleFilePath()
C <- readChromatogram(rawfile, mass = 445.1181, tol = 10)
plot(C[[1]])
}
\author{
Tobias Kockmann, 2020.
}
