% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{plot.rawrrChromatogramSet}
\alias{plot.rawrrChromatogramSet}
\title{Plot \code{rawrrChromatogramSet} objects}
\usage{
\method{plot}{rawrrChromatogramSet}(x, diagnostic = FALSE, ...)
}
\arguments{
\item{x}{A \code{rawrrChromatogramSet} object to be plotted.}

\item{diagnostic}{Show diagnostic legend?}

\item{...}{Passes additional arguments.}
}
\description{
Plot \code{rawrrChromatogramSet} objects
}
\author{
Tobias Kockmann, 2020.
}
