% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{plot.rawRspectrum}
\alias{plot.rawRspectrum}
\title{Basic plotting function for instances of \code{rawRspectrum}}
\usage{
Is usually called by method dispatch.
}
\arguments{
\item{x}{A \code{rawRspectrum} object}

\item{relative}{If set to \code{TRUE} enforces plotting of relative
intensities rather than absolute.}

\item{...}{function passes arbitrary additional arguments.}
}
\description{
\code{plot.rawRspectrum} is a low level function that calls
\code{base::plot} for plotting \code{rawRspectrum} objects. It passes all
additional arguments to \code{plot()}
}
\author{
Tobias Kockmann, 2020
}
