% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{print.rawrrSpectrum}
\alias{print.rawrrSpectrum}
\title{Print method imitate the look and feel of Thermo Fisher Scientific FreeStyle's output}
\usage{
\method{print}{rawrrSpectrum}(x, ...)
}
\arguments{
\item{x}{an \code{rawrrSpectrum} object.}

\item{\ldots}{Arguments to be passed to methods.}
}
\value{
This function creates a print message.
}
\description{
Print method imitate the look and feel of Thermo Fisher Scientific FreeStyle's output
}
\author{
Christian Panse and Tobias Kockmann, 2020.
}
