% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{print.rawRspectrum}
\alias{print.rawRspectrum}
\title{Basic print function faking the look and feel of freestyle's output}
\usage{
\method{print}{rawRspectrum}(x, ...)
}
\arguments{
\item{x}{an \code{rawRspectrum} object.}

\item{\ldots}{Arguments to be passed to methods.}
}
\description{
Basic print function faking the look and feel of freestyle's output
}
\author{
Christian Panse and Tobias Kockmann, 2020.
}
