% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\docType{package}
\name{rawrr-package}
\alias{rawrr-package}
\title{rawrr: Direct Access to Orbitrap Data and Beyond}
\description{
This package wraps the functionality of the RawFileReader .NET assembly. Within the R environment, spectra and chromatograms are represented by S3 objects. The package provides basic functions to download and install the required third-party libraries. The package is developed, tested, and used at the Functional Genomics Center Zurich, Switzerland.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fgcz/rawrr/}
  \item Report bugs at \url{https://github.com/fgcz/rawrr/issues}
}

}
\author{
\strong{Maintainer}: Christian Panse \email{cp@fgcz.ethz.ch} (\href{https://orcid.org/0000-0003-1975-3064}{ORCID})

Authors:
\itemize{
  \item Tobias Kockmann \email{tobias.kockmann@fgcz.ethz.ch} (\href{https://orcid.org/0000-0002-1847-885X}{ORCID})
}

}
\keyword{internal}
