% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{rawrrAssemblyPath}
\alias{rawrrAssemblyPath}
\title{Derives the path where all .NET assemblies are stored.}
\usage{
rawrrAssemblyPath()
}
\value{
path
}
\description{
Derives the path where all .NET assemblies are stored.
}
\examples{
rawrrAssemblyPath()
}
\seealso{
\code{installRawFileReaderDLLs} and \code{installRawrrExe}
}
