% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{rawrrSpectrum}
\alias{rawrrSpectrum}
\title{Create \code{rawrrSpectrum} objects}
\usage{
rawrrSpectrum(sim = character())
}
\arguments{
\item{sim}{Either \code{example_1} or \code{TESTPEPTIDE}}
}
\value{
Function returns a validated \code{rawrrSpectrum} object
}
\description{
High-level constructor for instances of class
\code{rawrrSpectrum}, also named helper function. Currently, mainly to support
testing and for demonstration.
}
\examples{
plot(rawrrSpectrum(sim = "TESTPEPTIDE"))
rawrrSpectrum(sim = "example_1")

}
\author{
Tobias Kockmann, 2020.
}
