% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{rawRspectrum}
\alias{rawRspectrum}
\title{Create \code{rawRspectrum} objects}
\usage{
rawRspectrum(sim = character())
}
\arguments{
\item{sim}{Either \code{example_1} or \code{TESTPEPTIDE}}
}
\value{
Function returns a validated \code{rawRspectrum} object
}
\description{
High-level constructor for instances of class
\code{rawRspectrum}, also named helper function. Currently, mainly to support
testing and for demonstration.
}
\examples{

plot(rawRspectrum(sim = "TESTPEPTIDE"))
rawRspectrum(sim = "example_1")

}
\author{
Tobias Kockmann, 2020.
}
