% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{readTrailer}
\alias{readTrailer}
\title{Read and extract scan trailer from TFS raw files.}
\usage{
readTrailer(rawfile, label = NULL)
}
\arguments{
\item{rawfile}{the name of the raw file containing the mass spectrometry data from the Thermo Fisher Scientific instrument.}

\item{label}{if NULL; the function scans for all available labels.}
}
\value{
an vector of trailers or values of a given trailer. Of note,
the values are usually returned as a character.
}
\description{
Read and extract scan trailer from TFS raw files.
}
\examples{
rawrr::sampleFilePath() |> rawrr:::readTrailer()
rawrr::sampleFilePath() |> rawrr:::readTrailer("AGC:") |> head()
}
