% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{sample}
\alias{sample}
\alias{sample.raw}
\title{{\code{sample.raw}}}
\usage{
sample()
}
\value{
filepath of the sample.raw location
}
\description{
Thermo Fisher Scientific Q Exactive HF-X raw file
of size 1.5M bytes and checksum
\code{MD5 (sample.raw) = fe67058456c79af7442316c474d20e96}.
}
\details{
The binary example file sample.raw contains 574 fourier-transformed orbi trap
spectra (FTMS) recorded on a Thermo Fisher Scientific Q Exactive HF-X. The
mass spectrometer was operated in line with a nano electrospray source (NSI)
in positive mode (+). All spectra were written to disk after applying
centroiding (c) and lock mass correction. Additional raw data for
demonstration and extended testing is available through the
\strong{Lions love raw meat!}
}
\examples{
sample()
}
\seealso{
\href{tartare package}{https://bioconductor.org/packages/tartare/}.
}
\author{
Tobias Kockmann 2018
}
