% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{summary.rawrrChromatogram}
\alias{summary.rawrrChromatogram}
\title{Text summary of chromatogram}
\usage{
\method{summary}{rawrrChromatogram}(object, ...)
}
\arguments{
\item{object}{A \code{rawrrChromatogram} object}

\item{...}{Function passes additional arguments.}
}
\value{
A \code{rawrrChromatogram} object
}
\description{
Text summary of chromatogram
}
\examples{
C <- readChromatogram(rawfile = sampleFilePath(),
mass = c(445.1181, 519.1367))
summary(C[[1]])
summary(C[[2]])
}
