% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{summary.rawrrSpectrum}
\alias{summary.rawrrSpectrum}
\title{Basic summary function}
\usage{
\method{summary}{rawrrSpectrum}(object, ...)
}
\arguments{
\item{object}{an \code{rawrrSpectrum} object.}

\item{\ldots}{Arguments to be passed to methods.}
}
\description{
Basic summary function
}
\author{
Christian Panse and Tobias Kockmann, 2020.
}
