% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{summary.rawRspectrum}
\alias{summary.rawRspectrum}
\title{Basic summary function}
\usage{
\method{summary}{rawRspectrum}(object, ...)
}
\arguments{
\item{object}{an \code{rawRspectrum} object.}

\item{\ldots}{Arguments to be passed to methods.}
}
\description{
Basic summary function
}
\author{
Christian Panse & Tobias Kockmann, 2020
}
