% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorFunctions.R
\name{tic}
\alias{tic}
\title{Total ion current of a spectrum}
\usage{
tic(x)
}
\arguments{
\item{x}{A rawrrSpectrum object}
}
\value{
A double vector of length one.
}
\description{
Total ion current of a spectrum
}
\examples{
S <- readSpectrum(rawfile = sampleFilePath(), 1)
tic(S[[1]])
}
