% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_net.R
\name{user_net}
\alias{user_net}
\title{User Network}
\usage{
user_net(df)
}
\arguments{
\item{df}{a data frame, as produced by the add_threads() function}
}
\value{
an edge list, each tie means that the user responded to a comment made by another
}
\description{
User Network
}
