% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{validate_rawrrIndex}
\alias{validate_rawrrIndex}
\title{Validate output of the readIndex function}
\usage{
validate_rawrrIndex(x)
}
\arguments{
\item{x}{object to be validated.}
}
\value{
Validated \code{data.frame} of \code{readIndex} object
}
\description{
Checks the validity of an \code{readIndex} returned object.
}
\examples{
Idx <- readIndex(sampleFilePath())
validate_rawrrIndex(Idx)
}
\author{
Tobias Kockmann and Christian Panse, 2020-12-09.
}
