% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{validate_rawrrSpectrum}
\alias{validate_rawrrSpectrum}
\title{Validate instance of class rawrrSpectrum}
\usage{
validate_rawrrSpectrum(x)
}
\arguments{
\item{x}{object to be validated.}
}
\value{
Validated \code{rawrrSpectrum} object
}
\description{
Checks the validity of \code{rawrrSpectrum} object attributes.
}
\author{
Tobias Kockmann and Christian Panse, 2020.
}
