% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{auc.rawrrChromatogram}
\alias{auc.rawrrChromatogram}
\title{Example snippet for deriving an AUC}
\usage{
auc.rawrrChromatogram(x)
}
\arguments{
\item{x}{an rawrrChromatogram}
}
\value{
A numeric value.
}
\description{
Example snippet for deriving an AUC
}
