% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{buildRawrrExe}
\alias{buildRawrrExe}
\title{Build \code{rawrr.exe} console application.}
\usage{
buildRawrrExe()
}
\value{
the return value of the system2 command.
}
\description{
builds \code{rawrr.exe} file from C# source code requiring
.NET SDK. The console application \code{rawrr.exe}
is used by the package's reader functions through a \link{system2} call
or a \link{textConnection}.

To use this function, ensure that the local RawFileReader NuGet packages 
are added to the NuGet source list. You can accomplish this by 
downloading the necessary packages with 
\code{rawrr:::.downloadNupkgs()} and subsequently running 
\code{rawrr:::.addNupkgSource()}.
}
\references{
\itemize{
  \item{\url{https://www.mono-project.com/docs/advanced/assemblies-and-the-gac/}, 2020}
  \item{\url{https://planetorbitrap.com/rawfilereader}, 2020}
  \item{\url{https://github.com/thermofisherlsms/RawFileReader/}, 2024}
  \item{\url{https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/compiler-options/advanced}}
  \item{\doi{10.1021/acs.jproteome.0c00866}}
}
}
\seealso{
\link{installRawrrExe}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>, 2021, 2024
}
