% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{buildRawrrExe}
\alias{buildRawrrExe}
\title{Build \code{rawrr.exe} console application \code{rawrr.exe}}
\usage{
buildRawrrExe()
}
\value{
the return value of the system2 command.
}
\description{
builds \code{rawrr.exe} file from C# source code requiring 
xbuild or msbuild tools.
}
\details{
The console application \code{rawrr.exe} is used by the package's reader functions through a \link{system2} call.
}
\references{
\itemize{
  \item{\url{https://www.mono-project.com/docs/advanced/assemblies-and-the-gac/}}
  \item{\url{https://planetorbitrap.com/rawfilereader}}
  \item{\url{https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/compiler-options/advanced}}
}
}
\seealso{
\link{installRawrrExe} and \link{installRawFileReaderDLLs}
}
\author{
Christian Panse <cp@fgcz.ethz.ch>, 2021
}
