% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{.benchmark}
\alias{.benchmark}
\title{f <- "/Users/cp/Library/Caches/org.R-project.R/R/ExperimentHub/46314c3933e2_4590.raw"}
\usage{
.benchmark(f)
}
\description{
f <- "/Users/cp/Library/Caches/org.R-project.R/R/ExperimentHub/46314c3933e2_4590.raw"
}
