% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_keyword_urls}
\alias{download_keyword_urls}
\title{Download urls by subreddit}
\usage{
download_keyword_urls(keyword, sub = "all", sort_by = "top")
}
\arguments{
\item{keyword}{a search query, for more information building a search query go here: https://www.reddit.com/wiki/search}

\item{sub}{a subreddit, unless you want to search in all subreddits, in which case the value is "all" (default)}

\item{sort_by}{Can be one of "top" (default), "relevance", "hot", "new", or "comments".}
}
\value{
a tibble
}
\description{
Download urls by subreddit
}
\examples{
\dontrun{
df <- download_keyword_urls(keyword = "reticulate", sub = "rstats")
}
}
