% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{extract_thread}
\alias{extract_thread}
\title{Extract a thread from a path}
\usage{
extract_thread(path)
}
\arguments{
\item{path}{a character string that starts with /r/, which you obtain using either the download_sub_urls() or the download_keyword_urls() functions}
}
\value{
A list with two tibbles: 1. information on nodes and 2. an edge list
}
\description{
Extract a thread from a path
}
\examples{
\dontrun{
df <- extract_thread("/r/rstats/comments/f5bxyk/r_will_use_stringsasfactors_false_by_default_in/")
}
}
