% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praw.R
\name{init_reddit}
\alias{init_reddit}
\title{Initialize Reddit instance}
\usage{
init_reddit(username, password, client_id, client_secret)
}
\arguments{
\item{username}{username}

\item{password}{password}

\item{client_id}{client_id}

\item{client_secret}{client_secret}
}
\description{
You must register for an application at https://www.reddit.com/prefs/apps/
Follow these steps https://github.com/reddit-archive/reddit/wiki/OAuth2-Quick-Start-Example#first-steps
}
