% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{installRawFileReaderDLLs}
\alias{installRawFileReaderDLLs}
\title{installRawFileReaderDLLs (deprecated)}
\usage{
installRawFileReaderDLLs()
}
\description{
Download and install the New RawFileReader from Thermo Fisher
Scientific .Net assemblies in the directory provided by
\code{rawrrAssemblyPath()}.
}
