% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotNetAssembly.R
\name{installRawrrExe}
\alias{installRawrrExe}
\alias{rawrr.exe}
\title{Download and install the \code{rawrr.exe} console application}
\usage{
installRawrrExe(
  sourceUrl = "https://github.com/fgcz/rawrr/releases/download/1.5.3/rawrr.1.5.3.exe",
  ...
)
}
\arguments{
\item{sourceUrl}{url of \code{rawrr.exe} assembly.}

\item{...}{other parameter for \code{download.file}.}
}
\value{
An integer code, 0 for success and non-zero for
failure. For the "wget" and "curl" methods this is the status code returned
by the external program.
}
\description{
downloads and installs the \code{rawrr.exe} .Net assembly in 
the directory provided by \code{rawrrAssemblyPath()}.
}
\details{
The console application \code{rawrr.exe} is used by the package's
reader functions through a \link{system2} call.
}
\references{
\doi{10.1021/acs.jproteome.0c00866}
}
\seealso{
\link{buildRawrrExe}
}
