% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{is.rawRchromatogram}
\alias{is.rawRchromatogram}
\title{Check if object is instance of class \code{rawRchromatogram}}
\usage{
is.rawRchromatogram(x)
}
\arguments{
\item{x}{The object to be tested.}
}
\value{
Boolean
}
\description{
Check if object is instance of class \code{rawRchromatogram}
}
\examples{
rawfile <- sample()
C <- readChromatogram(rawfile, mass = 445.1181, tol = 10)
is.rawRchromatogram(C[[1]])
}
\author{
Tobias Kockmann, 2020.
}
