% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{is.rawrrSpectrum}
\alias{is.rawrrSpectrum}
\title{Function to check if an object is an instance of class \code{rawrrSpectrum}}
\usage{
is.rawrrSpectrum(x)
}
\arguments{
\item{x}{any R object to be tested.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Function to check if an object is an instance of class \code{rawrrSpectrum}
}
\examples{
S <- rawrr::sampleFilePath() |> rawrr::readSpectrum(scan = 1:10)
rawrr::is.rawrrSpectrum(S[[1]])
}
