% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{is.rawrrSpectrum}
\alias{is.rawrrSpectrum}
\title{Test if object is instance of class \code{rawrrSpectrum}}
\usage{
is.rawrrSpectrum(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
TRUE or FALSE
}
\description{
Test if object is instance of class \code{rawrrSpectrum}
}
\examples{
rawfile <- sampleFilePath()
S <- readSpectrum(rawfile, scan = 1:10)
is.rawrrSpectrum(S[[1]])
}
