% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{print.rawrrSpectrum}
\alias{print.rawrrSpectrum}
\title{Basic print function faking the look and feel of freestyle's output}
\usage{
\method{print}{rawrrSpectrum}(x, ...)
}
\arguments{
\item{x}{an \code{rawrrSpectrum} object.}

\item{\ldots}{Arguments to be passed to methods.}
}
\description{
Basic print function faking the look and feel of freestyle's output
}
\author{
Christian Panse and Tobias Kockmann, 2020.
}
