% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{readChromatogram}
\alias{readChromatogram}
\title{Extracts Chromatograms}
\usage{
readChromatogram(
  rawfile,
  mass = NULL,
  tol = 10,
  filter = "ms",
  type = "xic",
  mono = if (Sys.info()["sysname"] \%in\% c("Darwin", "Linux")) TRUE else FALSE,
  exe = file.path(path.package(package = "rawR"), "exec", "rawR.exe")
)
}
\arguments{
\item{rawfile}{the file name.}

\item{mass}{a vector of mass values iff \code{type = 'xic'}.}

\item{tol}{mass tolerance in ppm iff \code{type = 'xic'}.}

\item{filter}{defines the scan filter, default is \code{filter="ms"} if a
wrong filter is set the function will return \code{NULL} and draws a warning.}

\item{type}{\code{c(xic, bpc, tic)} for extracted ion , base peak or
total ion chromatogram.}

\item{mono}{if the mono enviroment should be used.}

\item{exe}{the exe file user by mono.}
}
\value{
chromatogram object(s) containing of a vector of \code{times} and a
corresponding vector of \code{intensities}.
}
\description{
Extracts Chromatograms
}
\examples{

# Example 1: not meaning full but proof-of-concept
(rawfile <- file.path(path.package(package = 'rawR'), 'extdata', 'sample.raw'))

XIC <- readChromatogram(rawfile, mass=c(669.8381, 726.8357), tol=1000)
plot(XIC)

BPC <- readChromatogram(rawfile, type='bpc')
plot(BPC)

TIC <- readChromatogram(rawfile, type='tic')
plot(TIC)

# Example 2: extract iRT peptides
iRTpeptide <- c("LGGNEQVTR", "YILAGVENSK", "GTFIIDPGGVIR", "GTFIIDPAAVIR",
  "GAGSSEPVTGLDAK", "TPVISGGPYEYR", "VEATFGVDESNAK",
  "TPVITGAPYEYR", "DGLDAASYYAPVR", "ADVTPADFSEWSK",
  "LFLQFGAQGSPFLK")

# [2H+] 
if (require(protViz)){
     (mZ <- (parentIonMass(iRTpeptide) + 1.008) / 2)
  }else{
     message("consider installing  https://CRAN.R-project.org/package=protViz")
}

\dontrun{
# https://fgcz-ms.uzh.ch/p2692/Proteomics/QEXACTIVEHFX_1/tobiasko_20180220_scanSpeed/20180220_14_autoQC01.raw
# md5 = 00ffee77b82202200e5aec0522729f51

rawfile <- file.path(Sys.getenv('HOME'), "Downloads", "20180220_14_autoQC01.raw")
X <- readChromatogram(rawfile, mZ)
}

}
\references{
\itemize{
  \item{\url{https://doi.org/10.5281/zenodo.2640013}}
  \item{the R function 1st appeared in
    \url{https://doi.org/10.1021/acs.jproteome.8b00173}}
}
}
\seealso{
Thermo Fisher NewRawfileReader C# code snippets
\url{https://planetorbitrap.com/rawfilereader}.

\link[rawDiag]{readXIC}
}
\author{
Christian Trachsel, Tobias Kockmann and
Christian Panse <cp@fgz.ethz.ch> 2018, 2019, 2020.
}
