% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{readFileHeader}
\alias{readFileHeader}
\title{read file header Information}
\usage{
readFileHeader(
  rawfile,
  mono = if (Sys.info()["sysname"] \%in\% c("Darwin", "Linux")) TRUE else FALSE,
  exe = file.path(path.package(package = "rawR"), "exec", "rawR.exe"),
  mono_path = "",
  argv = "infoR",
  system2_call = TRUE,
  method = "thermo"
)
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scietific raw file}

\item{mono}{enviroment}

\item{exe}{path of the executable.}

\item{mono_path}{default.}

\item{argv}{arguments, default.}

\item{system2_call}{system2 call, default.}

\item{method}{instrument vendor}
}
\value{
a list object
}
\description{
The function extracts some meta information from a given rawfile.
The R code output is parsed by the function and a list object is returned.
}
\examples{
(rawfile <- file.path(path.package(package = 'rawR'), 'extdata',
  'sample.raw'))

M <- readFileHeader(rawfile)
}
\seealso{
Thermo Fisher NewRawfileReader C# code snippets
\url{https://planetorbitrap.com/rawfilereader}.
}
\author{
Tobias Kockmann and Christian Panse 2018, 2019, 2020
}
