% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{readFileHeader}
\alias{readFileHeader}
\title{read file header Information}
\usage{
readFileHeader(
  rawfile,
  mono = if (Sys.info()["sysname"] \%in\% c("Darwin", "Linux")) TRUE else FALSE,
  exe = .rawrrAssembly(),
  mono_path = "",
  argv = "infoR",
  system2_call = TRUE,
  method = "thermo"
)
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scietific raw file}

\item{mono}{enviroment}

\item{exe}{path of the executable.}

\item{mono_path}{default.}

\item{argv}{arguments, default.}

\item{system2_call}{system2 call, default.}

\item{method}{instrument vendor}
}
\value{
a list object containing the following entries: RAW file version,
Creation date, Operator, Number of instruments, Description,
Instrument model, Instrument name, Serial number, Software version,
Firmware version, Units, Mass resolution, Number of scans,
Number of ms2 scans, Scan range, Time range, Mass range,
Scan filter (first scan), Scan filter (last scan), Total number of filters,
Sample name, Sample id, Sample type, Sample comment, Sample vial,
Sample volume, Sample injection volume, Sample row number,
Sample dilution factor, or Sample barcode.
}
\description{
The function extracts meta information from a given rawfile.
}
\examples{
(rawfile <- file.path(path.package(package = 'rawrr'), 'extdata',
  'sample.raw'))

M <- readFileHeader(rawfile)
}
\references{
Thermo Fisher Scientific's NewRawfileReader C# code snippets
\url{https://planetorbitrap.com/rawfilereader}.
}
\author{
Tobias Kockmann and Christian Panse 2018, 2019, 2020.
}
