% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{readIndex}
\alias{readIndex}
\title{Read scan index}
\usage{
readIndex(rawfile, tmpdir = tempdir())
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scietific raw file.}

\item{tmpdir}{a non-empty character vector giving the directory name; default
uses \code{tempdir()}.}
}
\value{
returns a \code{data.frame} with the column names
scanType, rtinseconds, precursorMass, and charge of all spectra.
}
\description{
Read scan index
}
\examples{
rawfile <- sampleFilePath()

Idx <- readIndex(rawfile)
table(Idx$scanType)
plot(Idx$rtinseconds, Idx$precursorMass, col=as.factor(Idx$charge), pch=16)

table(Idx$MSOrder)


# given you have a raw file with depende

}
\author{
Tobias Kockmann and Christian Panse <cp@fgz.ethz.ch>, 2020
}
