% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{readSpectrum}
\alias{readSpectrum}
\alias{plot.rawRSpectrum}
\alias{rawRspectrum}
\alias{rawR}
\alias{sample.raw}
\title{Read a Set of Spectra}
\usage{
readSpectrum(rawfile, scan = NULL, tmpdir = tempdir(), validate = FALSE)
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scietific raw file.}

\item{scan}{a vector of requested scan numbers.}

\item{tmpdir}{a non-empty character vector giving the directory name; default
uses \code{tempdir()}.}

\item{validate}{boolean default is \code{FALSE}.}
}
\value{
a list of \code{spectrum} objects.
}
\description{
the function reads scan information, e.g., charge, mZ,
or intensity of a given set of scan numbers using a dot net interface and
the ThermoFisher NewRawFileReader libraries.
}
\details{
The binary example file sample.raw contains 574 fourier-transformed orbi trap
spectra (FTMS) recorded on a Thermo Fisher Scientific Q Exactive HF-X. The
mass spectrometer was operated in line with a nano electrospray source (NSI)
in positive mode (+). All spectra were written to disk after applying
centroiding (c) and lock mass correction. Additional raw data for
demonstration and extended testing is available through the
\href{tartare package}{https://bioconductor.org/packages/tartare/}.
Lions love raw meat!
}
\references{
\itemize{
\item{Thermo Fisher NewRawfileReader C# code snippets
\url{https://planetorbitrap.com/rawfilereader}}.
\item{\url{https://doi.org/10.5281/zenodo.2640013}}
\item{the R function 1st appeared in 
\url{https://doi.org/10.1021/acs.jproteome.8b00173}.}
}
}
\seealso{
\link[rawDiag]{readScans}
}
\author{
Tobias Kockmann and Christian Panse <cp@fgz.ethz.ch> 2018, 2019, 2020
}
