% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{readSpectrum}
\alias{readSpectrum}
\alias{plot.rawRSpectrum}
\title{Read a Set of Spectra}
\usage{
readSpectrum(rawfile, scans, tmpdir = tempdir())
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scietific raw file.}

\item{scans}{a vector of requested scan numbers.}

\item{tmpdir}{a non-empty character vector giving the directory name; default
uses \code{tempdir()}.}
}
\value{
a list of \code{spectrum} objects.
}
\description{
the function reads scan information, e.g., charge, mZ,
or intensity of a given set of scan numbers using a dot net interface and
the ThermoFisher NewRawFileReader libraries.
}
\examples{
(rawfile <- file.path(path.package(package = 'rawR'), 'extdata',
  'sample.raw'))

S <- readSpectrum(rawfile, scans = 1:9)

S[[1]]

names(S[[1]])

plot.spectrum <- function(x, ...){
  plot(x$mZ, x$intensity, type='h')
  labels <- na.omit(lapply(x, function(y){if (length(y)==1){y}else{NA}}))
  legend("topright", paste(names(labels), labels, sep=": "), ...)
 }
 
}
\references{
\itemize{
\item{\url{https://doi.org/10.5281/zenodo.2640013}}
\item{the R function 1st appeared in
\url{https://doi.org/10.1021/acs.jproteome.8b00173}}
}
}
\seealso{
Thermo Fisher NewRawfileReader C# code snippets
\url{https://planetorbitrap.com/rawfilereader}.
}
\author{
Tobias Kockmann and Christian Panse <cp@fgz.ethz.ch> 2018, 2019, 2020
}
