% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{readSpectrum}
\alias{readSpectrum}
\title{readSpectrum of a raw file}
\usage{
readSpectrum(rawfile, scans, tmpdir = tempdir())
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scietific raw file which
the data are to be read from.}

\item{scans}{a vector of requested scan numbers.}

\item{tmpdir}{a non-empty character vector giving the directory name; default
uses \code{tempdir()}.}
}
\value{
peaklistSet, a nested list of \CRANpkg{protViz} peaklist objects.
}
\description{
the function reads scan information of a given set of scan
number using a dot net interface and the ThermoFisher NewRawFileReader 
libraries.
}
\examples{
(rawfile <- file.path(path.package(package = 'rawR'), 'extdata',
  'sample.raw'))

S <- readSpectrum(rawfile, scans = 1:9)

S[[1]]
 
}
\references{
\url{https://doi.org/10.5281/zenodo.2640013}
}
\author{
Christian Panse <cp@fgz.ethz.ch> 2018, 2019, 2020
}
