% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawrr.R
\name{readSpectrum}
\alias{readSpectrum}
\alias{rawrr}
\title{Reads spectral data from a raw file.}
\usage{
readSpectrum(rawfile, scan = NULL, tmpdir = tempdir(), validate = FALSE)
}
\arguments{
\item{rawfile}{the name of the Thermo Fisher Scientific raw file.}

\item{scan}{a vector of requested scan numbers.}

\item{tmpdir}{a non-empty character vector giving the directory name;
default uses \code{tempdir()}.}

\item{validate}{boolean default is \code{FALSE}.}
}
\value{
a nested list of \code{rawrrSpectrum} objects containing more than 50
values of scan information, e.g., the charge state, two vectors containing
the mZ and its corresponding intensity values or the AGC information,
mass calibration, ion optics \ldots
}
\description{
The function derives spectra of a given raw file and a given
vector of scan numbers.
}
\details{
All mass spectra are recorded by scanning detectors (mass analyzers)
that log signal intensities for ranges of mass to charge ratios (m/z), also
referred to as position. These recordings can be of continuous nature,
so-called profile data (p), or appear centroided (c) in case discrete
information (tuples of position and intensity values) are sufficient.
This heavily compacted data structure is often called a peak list.
In addition to signal intensities, a peak list can also cover additional
peak attributes like peak resolution (R), charge (z), or local noise
estimates. In short, the additional attributes further described the nature
of the original profile signal or help to group peak lists with respect to
their molecular nature or processing history. A well-known example is the
assignment of peaks to peak groups that constitute isotope patterns
(M, M+1, M+2, ...).
The names of objects encapsulated within \code{rawrrSpectrum} instances are
keys returned by the Thermo Fisher Scientific New RawFileReader API and the
corresponding values become data parts of the objects, typically vectors.
}
\examples{
(rawfile <- sampleFilePath())

S <- readSpectrum(rawfile, scan = 1:9)

S[[1]]

names(S[[1]])

plot(S[[1]])

\donttest{
# INPUT:
GAG <- "GAGSSEPVTGLDAK"

# fetch via FTP download link into your Downloads folder
# MSV000086542
# MD5 (20181113_010_autoQC01.raw) = a1f5df9627cf9e0d51ec1906776957ab
rawfile <- rawrr:::.download_20181113_010_autoQC01()

# list spectra metainformation
IDX <- readIndex(rawfile)

# determine precursor matches
S <- readSpectrum(rawfile,
  which(abs((1.008 + (protViz::parentIonMass(GAG) - 1.008) / 2) - IDX$precursorMass) < 0.001))

# query spectra with precursor matches
rv <- lapply(S, function(x){protViz::psm(GAG, x, plot=FALSE)})

# determine spectra indices having the  max number of hits hits
hit.max <- max(hits <- sapply(rv, function(x){sum(abs(x$mZ.Da.error) < 0.01)}))

# take the 1st one
idx <- which(hits == hit.max)[1]

# OUTPUT
rv <- protViz::peakplot(GAG,  (S[[idx]]), FUN=function(b,y){cbind(b=b, y=y)})
# https://www.proteomicsdb.org/use/
cat(paste(S[[idx]]$mZ[rv$idx], "\t", S[[idx]]$intensity[rv$idx]), sep = "\n")
}
}
\references{
\itemize{
  \item{C# code snippets of the NewRawfileReader library
    \url{https://planetorbitrap.com/rawfilereader}.}
    \item{\url{https://doi.org/10.1021/acs.jproteome.0c00866}}
  }
}
\seealso{
\url{https://massive.ucsd.edu/ProteoSAFe/dataset.jsp?accession=MSV000086542}
}
\author{
Tobias Kockmann and Christian Panse <cp@fgz.ethz.ch> 2018, 2019, 2020
}
