% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawR.R
\name{validate_rawRspectrum}
\alias{validate_rawRspectrum}
\title{Validate instance of class rawRSpectrum}
\usage{
validate_rawRspectrum(x)
}
\arguments{
\item{x}{object to be validated.}
}
\value{
Validated \code{rawRspectrum} object
}
\description{
Checks the validity of \code{rawRspectrum} object attributes.
}
